requirements.txt
python-dotenv
Django>=3.2,<4.0
djangorestframework
drf-spectacular